Paisa_app/paisa/backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10          # alembic env (sync)

# ── Cache / Queue ─────────────────────────────────────────────────────
redis[hiredis]==5.2.1
celery[redis]==5.4.0
kombu==5.4.2

# ── Auth & Security ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==43.0.3

# ── S3-compatible storage ─────────────────────────────────────────────
boto3==1.35.82
aioboto3==13.2.0

# ── HTTP / Push ───────────────────────────────────────────────────────
httpx==0.28.1
firebase-admin==6.6.0            # FCM push notifications

# ── Observability ─────────────────────────────────────────────────────
structlog==24.4.0
opentelemetry-api==1.28.4
opentelemetry-sdk==1.28.4
opentelemetry-instrumentation-fastapi==0.49b4
opentelemetry-instrumentation-sqlalchemy==0.49b4
prometheus-client==0.21.1

# ── Validation / Serialization ────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Testing ───────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1                    # async test client
factory-boy==3.3.1
faker==33.1.0

# ── Dev utilities ─────────────────────────────────────────────────────
ruff==0.8.4
mypy==1.13.0
pre-commit==4.0.1
